Recursion Lab
Write a recursive method to sum the values in an array of integers.  Create a file ArraySum.java and add the recursive method 
public int sumOfArray (Integer[] a,int index).
Note that ‘a’ is an array of type Integer that is specified in the driver file, and ‘index’ is an integer that shows which number in the array to sum next.  
Use the driver class ArraySumDriver.java to populate your array and demonstrate that your method works.  
Load the files ArraySum.java and ArraySumDriver.java to GitHub and take a screen shot.
Upload your ArraySum.java, ArraySumDriver.java and your GitHub screen shot to Blackboard.